refactor/client filters actions dispatch
